fix: 모집글 중복 삭제시 예외 처리를 위한 서비스 레이어 로직 추가

+++ api-server/src/main/java/com/lastone/apiserver/service/recruitment/RecruitmentServiceImpl.java
@@ -2,6 +2,7 @@ package com.lastone.apiserver.service.recruitment;
 
 import com.lastone.apiserver.exception.mypage.MemberNotFountException;
 import com.lastone.apiserver.exception.recruitment.IncorrectWriterException;
+import com.lastone.apiserver.exception.recruitment.RecruitmentAlreadyDeleteException;
 import com.lastone.apiserver.exception.recruitment.RecruitmentImgCountException;
 import com.lastone.apiserver.exception.recruitment.RecruitmentNotFoundException;
 import com.lastone.apiserver.service.s3.S3Service;
@@ -113,6 +114,9 @@ public class RecruitmentServiceImpl implements RecruitmentService {
     @Override
     public void deleteRecruitment(Long recruitmentId, Long memberId) {
         Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(RecruitmentNotFoundException::new);
+        if (recruitment.isDeleted()) {
+            throw new RecruitmentAlreadyDeleteException();
+        }
         validateWriterAndMember(recruitment.getMember().getId(), memberId);
         deleteImgFile(recruitment.getRecruitmentImgs());
         recruitment.delete();
